refactor(daemon): cleanup
